requirements.txt
fastapi
uvicorn
python-multipart
pydantic
openai
faiss-cpu
sentence-transformers
pdfplumber
python-docx
numpy
pytest
